Move metadata integrations
