+++ src/palace/manager/api/lanes.py
@@ -5,8 +5,6 @@ from sqlalchemy.orm import Session
 
 import palace.manager.core.classifier as genres
 from palace.manager.api.config import CannotLoadConfiguration, Configuration
-from palace.manager.api.metadata.novelist import NoveListAPI
-from palace.manager.api.metadata.nyt import NYTBestSellerAPI
 from palace.manager.core import classifier
 from palace.manager.core.classifier import (
     Classifier,
@@ -14,6 +12,8 @@ from palace.manager.core.classifier import (
     fiction_genres,
     nonfiction_genres,
 )
+from palace.manager.integration.metadata.novelist import NoveListAPI
+from palace.manager.integration.metadata.nyt import NYTBestSellerAPI
 from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.contributor import Contributor
 from palace.manager.sqlalchemy.model.datasource import DataSource
