+++ src/palace/manager/integration/metadata/base.py
