+++ src/palace/manager/integration/metadata/novelist.py
@@ -14,11 +14,14 @@ from sqlalchemy.engine import Row
 from sqlalchemy.orm import Session, aliased
 from sqlalchemy.sql import and_, join, or_, select
 
-from palace.manager.api.metadata.base import MetadataService, MetadataServiceSettings
 from palace.manager.core.config import CannotLoadConfiguration
 from palace.manager.data_layer.identifier import IdentifierData
 from palace.manager.integration.base import HasLibraryIntegrationConfiguration
 from palace.manager.integration.goals import Goals
+from palace.manager.integration.metadata.base import (
+    MetadataService,
+    MetadataServiceSettings,
+)
 from palace.manager.integration.settings import (
     BaseSettings,
     ConfigurationFormItem,
