+++ src/palace/manager/service/integration_registry/metadata.py
@@ -1,7 +1,7 @@
-from palace.manager.api.metadata.base import MetadataServiceType
-from palace.manager.api.metadata.novelist import NoveListAPI
-from palace.manager.api.metadata.nyt import NYTBestSellerAPI
 from palace.manager.integration.goals import Goals
+from palace.manager.integration.metadata.base import MetadataServiceType
+from palace.manager.integration.metadata.novelist import NoveListAPI
+from palace.manager.integration.metadata.nyt import NYTBestSellerAPI
 from palace.manager.service.integration_registry.base import IntegrationRegistry
