+++ tests/manager/celery/tasks/test_novelist.py
@@ -2,12 +2,15 @@ from unittest.mock import create_autospec, patch
 
 import pytest
 
-from palace.manager.api.metadata.novelist import NoveListAPI, NoveListApiSettings
 from palace.manager.celery.tasks.novelist import (
     update_novelists_by_library,
     update_novelists_for_all_libraries,
 )
 from palace.manager.integration.goals import Goals
+from palace.manager.integration.metadata.novelist import (
+    NoveListAPI,
+    NoveListApiSettings,
+)
 from palace.manager.service.logging.configuration import LogLevel
 from tests.fixtures.celery import CeleryFixture
 from tests.fixtures.database import DatabaseTransactionFixture
