+++ tests/manager/celery/tasks/test_nyt.py
@@ -3,8 +3,8 @@ from unittest.mock import create_autospec, patch
 
 import pytest
 
-from palace.manager.api.metadata.nyt import NYTBestSellerAPI, NYTBestSellerList
 from palace.manager.celery.tasks.nyt import update_nyt_best_sellers_lists
+from palace.manager.integration.metadata.nyt import NYTBestSellerAPI, NYTBestSellerList
 from tests.fixtures.celery import CeleryFixture
 from tests.fixtures.database import DatabaseTransactionFixture
