+++ tests/manager/integration/metadata/test_novelist.py
@@ -11,9 +11,12 @@ import pytest
 from pytest import MonkeyPatch
 
 from palace.manager.api.config import CannotLoadConfiguration
-from palace.manager.api.metadata.novelist import NoveListAPI, NoveListApiSettings
 from palace.manager.data_layer.identifier import IdentifierData
 from palace.manager.integration.goals import Goals
+from palace.manager.integration.metadata.novelist import (
+    NoveListAPI,
+    NoveListApiSettings,
+)
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.identifier import Identifier
 from palace.manager.sqlalchemy.model.resource import HttpResponseTuple, Representation
