+++ tests/manager/scripts/test_novelist.py
@@ -4,8 +4,11 @@ from unittest.mock import patch
 
 import pytest
 
-from palace.manager.api.metadata.novelist import NoveListAPI, NoveListApiSettings
 from palace.manager.integration.goals import Goals
+from palace.manager.integration.metadata.novelist import (
+    NoveListAPI,
+    NoveListApiSettings,
+)
 from palace.manager.scripts.novelist import NovelistSnapshotScript
 from palace.manager.service.logging.configuration import LogLevel
 from tests.fixtures.database import DatabaseTransactionFixture
